Fix all mypy/ruff errors and logic bugs before feature merge
mypy (13 errors → 0):
- analyze_metrics.py: rename CSV loop var from r to row to prevent
  type collision with the later analysis loop variable r
- ui/sim_panel.py:51: wrap 10.0 ** expr in float() to avoid Any return
- ui/fpga_board.py: add list[LED/Button/Switch] annotations on the
  if-branch assignments; rename sw,sh to scr_w,scr_h to prevent sw
  being fixed as int before the for-sw-in-switches loop
- sim_testbench.py: introduce _metrics_obj (SimMetrics) inside the
  if-block so .start() is called on a non-None type, then widen to
  SimMetrics | None for the rest of the function

ruff (1 error → 0):
- ui/fpga_board.py: add all 7 missing parameter descriptions to the
  __init__ docstring (D417)

Logic fixes:
- fpga_board.py: import get_font and call cache_clear() before each
  of the 3 pygame.quit() calls to prevent stale Font objects being
  returned after the next pygame.init()
- sim_testbench.py: fix _BASE_STEP_NS comment to say 383840/9596
  instead of the old pre-VHDL-clock values 9720/243

+++ ui/sim_panel.py
@@ -48,7 +48,7 @@ def _speed_to_frac(speed: float) -> float:
 
 def _frac_to_speed(frac: float) -> float:
     """Map a slider fraction in [0, 1] to a speed multiplier."""
-    return 10.0 ** (max(0.0, min(1.0, frac)) * _LOG_RANGE + _LOG_MIN)
+    return float(10.0 ** (max(0.0, min(1.0, frac)) * _LOG_RANGE + _LOG_MIN))
 
 
 def _fmt_hz(hz: float) -> str:
